api/requirements.txt
paddlepaddle==3.2.2
paddleocr
fastapi
uvicorn
python-multipart
pillow